Remove destination uri endpoint parameter
The original plan was to allow for a 'resouffle' option on the web
client. In this case, client would send uri of the already generated
playlist, and server would delete all those tracks and repopulate
with the new tracks.
But according to docs it seems that recommended 'delete playlist'
action is just to delete it, which makes more sense to do on client
side.