Add hacky fix for loading playlists directly from /playlists route
Note: This is a super hacky fix so that if a user loads the site at
/playlists, the user's playlists won't be fetched until the
handleApplicationStarted thunk has fetched an access token (if there
is a refreshToken stored in localStorage). This also required that I
set fetchingAccessToken and fetchingRefreshToken to true as part of
the auth reducer's initial state.playlists.